pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "showdown-optimizer-v2"
version = "0.1.0"
description = "DFS showdown optimizer/training pipelines (Pipeline A/B)"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "numpy>=1.26.0",
  "pandas>=2.2.0",
  "pydantic>=2.7.0",
  "pyyaml>=6.0.1",
  "typer>=0.12.3",
  "pyarrow>=16.0.0",
  "structlog>=24.1.0",
  "numba>=0.59.0",
  "scipy>=1.11.0",
  "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2.0",
  "ruff>=0.6.0",
  "black>=24.4.2",
  "isort>=5.13.2",
]

[project.scripts]
dfs-opt = "dfs_opt.cli.main:app"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.package-data]
dfs_opt = ["resources/*.yaml"]

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.isort]
profile = "black"
line_length = 100

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]
ignore = ["E501"]


